Adding phone company of the sending SMS number
While sharing some data using this object, we saw the need to add the phone company of the number sending the sms. 
With it we can make good local correlations and have an idea of flaws ocurring on phone number release by these companies.
Using web services like Truecaller, it's possible to enrich an analysis with this data.